Documentacao/requirements.txt
python-docx>=0.8.11
json 
os
zipfile
shutil
datetime
python-docx
